continuing of the refactor for expressCheckout container file tree.
