--- src/containers/checkout/component.imports.js
@@ -1,23 +1,0 @@
-export BreadCrumb from './components/BreadCrumb';
-export BillingAddress from './components/billingAddress/';
-export ShippingAddress from './components/shippingAddress/';
-export ShippingMethod from './components/shippingMethod';
-export CreditCardInfo from './components/creditCardInfo/';
-export ProductReview from './components/productReview/';
-export GrandTotal from './components/grandTotal/';
-export SubmitOrder from './components/submitOrder/';
-export NetworkStatus from './components/networkStatus';
-export CvnModal from './components/creditCardInfo/cvnModal';
-
-export FirstName from './components/shippingAddress/firstName';
-export LastName from './components/shippingAddress/lastName';
-export Email from './components/shippingAddress/email';
-export AddressLine from './components/shippingAddress/addressLine';
-export Country from './components/shippingAddress/country';
-export Prefecture from './components/shippingAddress/prefecture';
-export PostalCode from './components/shippingAddress/postalCode';
-export City from './components/shippingAddress/city';
-export PhoneNumber from './components/shippingAddress/phoneNumber';
-export Prefectures from './components/prefectureConstants';
-export Countries from './components/countryConstants';
-export States from './components/stateConstants';
