--- src/containers/checkout/orderSuccess/components/index.js
@@ -1,5 +1,0 @@
-export OrderHeader from './orderHeader';
-export ShipTo from './shipTo';
-export BillTo from './billTo';
-export OrderSummary from './orderSummary';
-export BreadCrumb from './BreadCrumb';
