+++ src/containers/orderSuccess/assets/graphql/index.js
@@ -1,0 +1,4 @@
+export {
+  FetchSagawa,
+  FetchMultipleProducts,
+} from './queries';
