+++ src/containers/orderSuccess/assets/graphql/queries.js
@@ -1,0 +1,4 @@
+export {
+  FetchSagawa,
+  FetchMultipleProducts,
+} from '../../../../graphql/queries';
