+++ src/containers/orderSuccess/assets/utils/index.js
@@ -1,0 +1,2 @@
+export zipArrays from '../../../../services/utils/zipArrays';
+export arrayDeepEquality from '../../../../services/utils/arrayDeepEquality';
