+++ src/containers/orderSuccess/components/index.js
@@ -1,0 +1,5 @@
+export OrderHeader from './OrderHeader';
+export ShipTo from './ShipTo';
+export BillTo from './BillTo';
+export OrderSummary from './OrderSummary';
+export BreadCrumb from './BreadCrumb';
